refactor participants migrations
