+++ protocols/migration/migration_reports_3_0_0_to_participant_1_0_0.py
@@ -8,12 +8,109 @@ from protocols.util import handle_avro_errors
 from protocols.migration.base_migration import BaseMigration, MigrationError
 
 
-class MigrateReports3ToParticipant1(BaseMigration):
+class MigrationReports3ToParticipant1(BaseMigration):
     """
     Any participant with empty labId in tumour or germline sample will fail to validate.
     """
     old_model = reports_3_0_0
     new_model = participant_1_0_0
+
+    def migrate_pedigree(self, pedigree, ready_for_analysis=True):
+        """
+
+        :type pedigree: participant_old.Pedigree
+        :rtype: participant_1_0_1.Pedigree
+        """
+        new_pedigree = self.convert_class(self.new_model.Pedigree, pedigree)
+        new_pedigree.versionControl = self.new_model.VersionControl()
+        new_pedigree.members = self.convert_collection(pedigree.participants, self.migrate_pedigree_member)
+        new_pedigree.analysisPanels = self.convert_collection(
+            pedigree.analysisPanels, self.migrate_analysis_panel, default=[])
+        new_pedigree.readyForAnalysis = ready_for_analysis
+        new_pedigree.familyId = pedigree.gelFamilyId
+        if new_pedigree.validate(new_pedigree.toJsonDict()):
+            return new_pedigree
+        else:
+            raise MigrationError('This model can not be converted')
+
+    def migrate_pedigree_member(self, member, sample_id_to_lab_sample_id=None):
+        """
+
+        :type member: participant_old.RDParticipant
+        :rtype: participant_1_0_1.PedigreeMember
+        """
+        new_pedigree_member = self.convert_class(self.new_model.PedigreeMember, member)
+        new_pedigree_member.participantId = member.gelId
+        new_pedigree_member.sex = self.migrate_enumerations('Sex', member.sex)
+        new_pedigree_member.lifeStatus = self.migrate_enumerations('LifeStatus', member.lifeStatus)
+        new_pedigree_member.adoptedStatus = self.migrate_enumerations('AdoptedStatus', member.adoptedStatus)
+        new_pedigree_member.affectionStatus = self.migrate_enumerations('AffectionStatus', member.affectionStatus)
+        new_pedigree_member.hpoTermList = self.convert_collection(member.hpoTermList, self.migrate_hpo_terms)
+        try:
+            new_pedigree_member.yearOfBirth = self.convert_string_to_integer(member.yearOfBirth)
+        except MigrationError:
+            new_pedigree_member.yearOfBirth = None
+            logging.warning("We are losing the year of birth as it cannot be converted into an integer")
+
+        new_pedigree_member.samples = []
+        if member.samples is not None:
+            for sample in member.samples:
+                if sample_id_to_lab_sample_id is not None and isinstance(sample_id_to_lab_sample_id, dict):
+                    lab_sample_id = self.convert_string_to_integer(
+                        sample_id_to_lab_sample_id.get(sample, -1), default_value=-1)
+                else:
+                    lab_sample_id = -1
+                new_pedigree_member.samples.append(self.new_model.Sample(
+                    sampleId=sample,
+                    labSampleId=lab_sample_id
+                ))
+        if new_pedigree_member.validate(new_pedigree_member.toJsonDict()):
+            return new_pedigree_member
+        else:
+            raise Exception('This model can not be converted')
+
+    def migrate_enumerations(self, etype, value):
+        if etype in ['LifeStatus', 'AffectionStatus']:
+            if etype == 'AffectionStatus' and value == self.old_model.AffectionStatus.unknown:
+                return self.new_model.AffectionStatus.UNCERTAIN
+            else:
+                return value.upper()
+        elif etype == 'Sex':
+            return {'male': 'MALE', 'female': 'FEMALE', 'unknown': 'UNKNOWN', 'undetermined': 'UNKNOWN'}.get(value)
+        elif etype == 'AdoptedStatus':
+            return {'not_adopted': 'notadopted', 'adoptedin': 'adoptedin', 'adoptedout': 'adoptedout'}.get(value)
+        elif etype == 'termPresence':
+            return {True: 'yes', False: 'no', None: 'unknown'}.get(value)
+        else:
+            raise NotImplementedError(etype + ' is not a valid enumeration type or is not implemented')
+
+    def migrate_hpo_terms(self, hpo_term):
+        """
+
+        :type hpo_term: participant_old.HpoTerm
+        :rtype: participant_1_0_1.HpoTerm
+        """
+        new_hpo = self.convert_class(self.new_model.HpoTerm, hpo_term)
+        new_hpo.termPresence = self.migrate_enumerations('termPresence', hpo_term.termPresence)
+        if new_hpo.validate(new_hpo.toJsonDict()):
+            return new_hpo
+        else:
+            raise Exception('This model can not be converted')
+
+    def migrate_analysis_panel(self, analysis_panel):
+        """
+
+        :type analysis_panel: participant_old.AnalysisPanel
+        :rtype: participant_1_0_1.AnalysisPanel
+        """
+
+        new_analysis_panel = self.convert_class(self.new_model.AnalysisPanel, analysis_panel)
+        new_analysis_panel.multipleGeneticOrigins = ''
+        new_analysis_panel.reviewOutcome = ''
+        if new_analysis_panel.validate(new_analysis_panel.toJsonDict()):
+            return new_analysis_panel
+        else:
+            raise Exception('This model can not be converted')
 
     def migrate_cancer_participant(self, old_cancer_participant):
         """
